requirements.txt
certifi==2019.9.11
cffi==1.13.2
cryptography==2.8
future==0.18.2
Pillow==6.2.1
pycparser==2.19
python-telegram-bot==12.2.0
six==1.13.0
tornado==6.0.3
